requirements.txt
fastapi==0.110.*
bs4==0.0.2
uvicorn
spacy>=3.0.0,<4.0.0
pyspellchecker==0.5.*